pyproject.toml
[project]
name = "litemcp"
version = "0.1.1"
description = "A minimal, lightweight client designed to simplify SDK (e.g., LangChain, Agent SDK) adoption into MCP."
authors = [{ name = "myan", email = "myan@redhat.com" }]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
  "mcp (>=1.4.1,<2.0.0)",
  "rich (>=13.9.4,<14.0.0)",
  "openai-agents (>=0.0.6,<0.0.7)",
  "commentjson (>=0.9.0,<0.10.0)",
  "langchain-core (>=0.3.47,<0.4.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"
openai = "^1.68.0"
langchain-openai = "^0.3.9"
